Allowed Cors *

+++ Backend/internal/controllers/service_ip_controller.go
@@ -7,46 +7,73 @@ import (
 	"backend/internal/models"
 
 	"github.com/gin-gonic/gin"
+	"github.com/google/uuid"
 )
 
-// Add a new allowed IP
 func AddAllowedIp(c *gin.Context) {
-	var input models.AllowedIp
+	if c.GetString("role") != "super_admin" {
+		c.JSON(http.StatusForbidden, gin.H{"error": "insufficient privileges"})
+		return
+	}
+
+	var input struct {
+		Service string ` binding:"require,max=40" json:"service"`
+		Ip      string ` binding:"required,ip" json:"ip"`
+	}
+
 	if err := c.ShouldBindJSON(&input); err != nil {
 		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
 		return
 	}
-	if err := config.DB.Create(&input).Error; err != nil {
+
+	allowed_ip := models.AllowedIp{
+		ID:      uuid.New().String(),
+		Service: input.Service,
+		Ip:      input.Ip,
+	}
+	if err := config.DB.Create(&allowed_ip).Error; err != nil {
 		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to add IP"})
 		return
 	}
-	c.JSON(http.StatusOK, input)
+	c.JSON(http.StatusOK, allowed_ip)
 }
 
-// Update an existing allowed IP
 func UpdateAllowedIp(c *gin.Context) {
+
+	if c.GetString("role") != "super_admin" {
+		c.JSON(http.StatusForbidden, gin.H{"error": "insufficient privileges"})
+		return
+	}
+
 	var ip models.AllowedIp
 	id := c.Param("id")
-	if err := config.DB.First(&ip, id).Error; err != nil {
+	if err := config.DB.Where("id = ?", id).First(&ip).Error; err != nil {
 		c.JSON(http.StatusNotFound, gin.H{"error": "IP not found"})
 		return
 	}
 
-	var input models.AllowedIp
+	var input struct {
+		Ip string ` binding:"required,ip" json:"ip"`
+	}
+
 	if err := c.ShouldBindJSON(&input); err != nil {
 		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
 		return
 	}
 
-	ip.Service = input.Service
 	ip.Ip = input.Ip
 	config.DB.Save(&ip)
 
 	c.JSON(http.StatusOK, ip)
 }
 
-// Delete an allowed IP
 func DeleteAllowedIp(c *gin.Context) {
+
+	if c.GetString("role") != "super_admin" {
+		c.JSON(http.StatusForbidden, gin.H{"error": "insufficient privileges"})
+		return
+	}
+
 	id := c.Param("id")
 	if err := config.DB.Delete(&models.AllowedIp{}, id).Error; err != nil {
 		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to delete IP"})
@@ -55,8 +82,13 @@ func DeleteAllowedIp(c *gin.Context) {
 	c.JSON(http.StatusOK, gin.H{"message": "IP deleted"})
 }
 
-// Get all allowed IPs or filter by service
 func GetAllowedIps(c *gin.Context) {
+
+	if c.GetString("role") != "super_admin" {
+		c.JSON(http.StatusForbidden, gin.H{"error": "insufficient privileges"})
+		return
+	}
+
 	service := c.Query("service")
 	var ips []models.AllowedIp
 	query := config.DB
@@ -68,19 +100,4 @@ func GetAllowedIps(c *gin.Context) {
 		return
 	}
 	c.JSON(http.StatusOK, ips)
-}
-
-// Check if IP is allowed for a given service
-func CheckIfAllowed(c *gin.Context) {
-	service := c.Query("service")
-	ip := c.Query("ip")
-
-	var count int64
-	config.DB.Model(&models.AllowedIp{}).Where("service = ? AND ip = ?", service, ip).Count(&count)
-
-	if count > 0 {
-		c.JSON(http.StatusOK, gin.H{"allowed": true})
-	} else {
-		c.JSON(http.StatusOK, gin.H{"allowed": false})
-	}
 }
